re-organizing everything
I really have no idea what I'm doing
