+++ keyboards/ogurec/config.h
@@ -21,8 +21,8 @@
 #define MANUFACTURER DrHigsby
 #define PRODUCT ogurec
 #define VENDOR_ID 0x0007
-#define PRODUCT_ID 0x0421
-#define DEVICE_VER 0x0420
+#define PRODUCT_ID 0x0002
+#define DEVICE_VER 0x0001
 
 /* key matrix size */
 #define MATRIX_ROWS 3
