+++ keyboards/ogurec/keymaps/dack/config.h
@@ -14,6 +14,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #pragma once
-
-#define COMBO_COUNT 7
-#define COMBO_TERM 40
+#define IGNORE_MOD_TAP_INTERRUPT
+#define TAPPING_TERM 230
+#define TAPPING_TERM_PER_KEY
