+++ keyboards/ogurec/keymaps/dack/readme.md
@@ -1,0 +1,3 @@
+# Dack's Ogurec Keymap
+
+This is Dack's personal ogurec keymap and it is for left-sided PM placement
