+++ keyboards/ogurec/keymaps/default/readme.md
@@ -1,3 +1,1 @@
-# Default Ogurec Keymap
-
-The default ogurec keymap is for left-sided PM placement
+# default ogurec keymap
