--- keyboards/ogurec/keymaps/right/rules.mk
@@ -1,1 +1,0 @@
-COMBO_ENABLE = yes
