+++ keyboards/ogurec/left_pm/readme.md
@@ -2,7 +2,11 @@
 
 Make example for this keyboard (after setting up your build environment):
 
-    make ogurec:left
+    make ogurec:default
+
+Flashing example for this keyboard:
+
+    make ogurec:default:flash
 
 To place the keyboard into bootloader mode, solder a reset switch onto the PCB and press it OR short the GND and RST pins on the pro micro.
